Implementation of Continuous Integration

+++ cronos_block/configuration/ShockTubeSod/modules.C
@@ -19,7 +19,7 @@ void Environment::setType(Data &gdata) {
 
 
 #if (NON_LINEAR_GRID == CRONOS_ON)
-void Grid::set_UserGridFunction(int dir, REAL beg, REAL end, REAL len) {
+void Grid::set_UserGridFunction(int dir, double beg, double end, double len) {
 //	myNonlinGrid[dir] = new GridFunction_Sin();
 	myNonlinGrid[dir] = new GridFunction_modEXP(beg,end,len);
   // myNonlinGrid[dir] = new GridFunction_Lin();
